requirements.txt
alabaster==0.7.10
amqp==2.2.1
Babel==2.5.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.3
django-rest-swagger==2.1.2
django-simple-history==1.9.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
docutils==0.14
gunicorn==19.7.1
idna==2.6
imagesize==0.7.1
itypes==1.1.0
Jinja2==2.9.6
kombu==4.1.0
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.2
Pillow==4.2.1
psycopg2==2.7.3
Pygments==2.2.0
PyJWT==1.5.2
pytz==2017.2
redis==2.10.6
requests==2.18.4
simplejson==3.11.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.15
vine==1.1.4
